+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change log for sw.allotmentclub
 4.12.3 (unreleased)
 ===================
 
-- Nothing changed yet.
+- FinTS-Integration now logs an error if a TAN is required.
 
 
 4.12.2 (2020-05-28)